lead-system/requirements.txt
playwright
beautifulsoup4
psycopg2-binary
python-dotenv
requests
sib-api-v3-sdk
